feat(list-campus): moved MainActivity.kt to presentation package and set as AndroidEntryPoint by hilt
