+++ app/src/main/java/com/byansanur/campuslist/presentation/MainActivity.kt
@@ -1,0 +1,22 @@
+package com.byansanur.campuslist.presentation
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import com.byansanur.campuslist.presentation.navigation.Navigation
+import com.byansanur.campuslist.ui.theme.CampusListTheme
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            CampusListTheme {
+                Navigation()
+            }
+        }
+    }
+}
